# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bare Metal Lifecycle Platform MVP API
  version: 0.1.0
  description: Unified bare metal lifecycle management platform MVP API draft.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/User' }
...
    LabValidationRunDetail:
      type: object
      properties:
        run: { $ref: '#/components/schemas/LabValidationRunSummary' }
        results:
          type: array
          items: { $ref: '#/components/schemas/LabValidationRunResult' }
...
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string, enum: [admin, operator, viewer] }
...
    MetadataNetwork:
      type: object
      properties:
        interfaces:
          type: array
          items: { $ref: '#/components/schemas/MetadataNetworkInterface' }
    MetadataSSHKeys:
      type: object
      properties:
        keys:
          type: array
          items: { type: string }
...
paths:
...
  /boot/ipxe:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Render server-specific iPXE script and auto-discover unknown MACs.
      description: Records BootEvent source as http_ipxe only when the client IP is inside an enabled deployment network CIDR; HTTP iPXE requests from outside deployment networks, or requests with X-Lab-Validation-Probe set to 1/true, are recorded as api_event for troubleshooting and do not prove physical PXE.
      parameters:
        - { name: mac, in: query, schema: { type: string } }
        - { name: arch, in: query, schema: { type: string } }
        - { name: firmware, in: query, schema: { type: string } }
      responses:
        '200': { description: iPXE script. }
...
  /metadata/instance-id:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata instance id by client IP.
      responses:
        '200': { description: Instance id text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Client IP did not match a server. }
  /metadata/hostname:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata hostname by client IP.
      responses:
        '200': { description: Hostname text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Client IP did not match a server. }
  /metadata/network:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata network JSON by client IP.
      responses:
        '200':
          description: Network metadata JSON.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetadataNetwork' }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Client IP did not match a server. }
  /metadata/ssh-keys:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return SSH public keys by client IP.
      responses:
        '200':
          description: SSH public keys JSON.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetadataSSHKeys' }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Client IP did not match a server. }
  /metadata/userdata:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return rendered userdata by client IP.
      responses:
        '200': { description: Userdata text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Client IP did not match a server. }
  /userdata:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Compatibility alias for client-IP userdata.
      responses:
        '200': { description: Userdata text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Client IP did not match a server. }
  /metadata/by-server/{id}/instance-id:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata instance id for a server.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Instance id text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
  /metadata/by-server/{id}/hostname:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata hostname for a server.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Hostname text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
  /metadata/by-server/{id}/network:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata network JSON for a server.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Network metadata JSON.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetadataNetwork' }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
  /metadata/by-server/{id}/ssh-keys:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return SSH public keys for a server metadata context.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: SSH public keys JSON.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetadataSSHKeys' }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
  /metadata/by-server/{id}/userdata:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return rendered userdata for a server.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Userdata text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
  /userdata/by-server/{id}:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Compatibility alias for server userdata.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Userdata text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
  /metadata/by-token/{token}/instance-id:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata instance id using a deployment metadata token.
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Instance id text. }
        '401': { description: Metadata token expired. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Metadata token not found. }
  /metadata/by-token/{token}/hostname:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata hostname using a deployment metadata token.
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Hostname text. }
        '401': { description: Metadata token expired. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Metadata token not found. }
  /metadata/by-token/{token}/network:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata network JSON using a deployment metadata token.
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Network metadata JSON.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetadataNetwork' }
        '401': { description: Metadata token expired. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Metadata token not found. }
  /metadata/by-token/{token}/ssh-keys:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return SSH public keys using a deployment metadata token.
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: SSH public keys JSON.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetadataSSHKeys' }
        '401': { description: Metadata token expired. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Metadata token not found. }
  /metadata/by-token/{token}/userdata:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return rendered userdata using a deployment metadata token.
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Userdata text. }
        '401': { description: Metadata token expired. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Metadata token not found. }
  /userdata/by-token/{token}:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Compatibility alias for token-based server userdata.
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Userdata text. }
        '401': { description: Metadata token expired. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Metadata token not found. }
  /metadata/by-mac/{mac}/{field}:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata by server primary MAC.
      description: Field must be one of instance-id, hostname, network, ssh-keys, or userdata.
      parameters:
        - { name: mac, in: path, required: true, schema: { type: string } }
        - { name: field, in: path, required: true, schema: { type: string, enum: [instance-id, hostname, network, ssh-keys, userdata] } }
      responses:
        '200': { description: Metadata field response. }
        '400': { description: Invalid MAC. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Server or field not found. }
  /userdata/by-mac/{mac}:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Compatibility alias for MAC-based userdata.
      parameters:
        - { name: mac, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Userdata text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Server not found. }
  /metadata/by-ip/{ip}/{field}:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata by server primary IP.
      description: Field must be one of instance-id, hostname, network, ssh-keys, or userdata.
      parameters:
        - { name: ip, in: path, required: true, schema: { type: string } }
        - { name: field, in: path, required: true, schema: { type: string, enum: [instance-id, hostname, network, ssh-keys, userdata] } }
      responses:
        '200': { description: Metadata field response. }
        '400': { description: Invalid IP. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Server or field not found. }
  /userdata/by-ip/{ip}:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Compatibility alias for IP-based userdata.
      parameters:
        - { name: ip, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Userdata text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Server not found. }
  /metadata/by-deployment/{id}/{field}:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Return metadata by deployment ID.
      description: Field must be one of instance-id, hostname, network, ssh-keys, or userdata.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: field, in: path, required: true, schema: { type: string, enum: [instance-id, hostname, network, ssh-keys, userdata] } }
      responses:
        '200': { description: Metadata field response. }
        '400': { description: Invalid deployment ID. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Deployment, server, or field not found. }
  /userdata/by-deployment/{id}:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Compatibility alias for deployment-based userdata.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Userdata text. }
        '403': { description: Metadata access is only allowed from enabled deployment networks when configured. }
        '404': { description: Deployment or server not found. }
  /images/{id}/file:
    get:
      servers:
        - url: http://localhost:8080
      security: []
      summary: Download enabled and verified image file.
      description: Serves only enabled, tested_passed images whose file_path resolves inside IMAGE_STORAGE_DIR. Supports HTTP Range requests.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: Range, in: header, schema: { type: string } }
      responses:
        '200': { description: Image file content. }
        '206': { description: Partial image file content. }
        '403': { description: Image is not enabled and verified, or file_path is outside image storage. }
        '404': { description: Image record or file not found. }
  /auth/login:
    post:
      security: []
      summary: Login with local account.
      description: Failed attempts are rate-limited by LOGIN_RATE_LIMIT_ATTEMPTS and LOGIN_RATE_LIMIT_WINDOW_SECONDS. 429 responses include Retry-After.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Login succeeded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { description: Invalid credentials. }
        '429': { description: Too many failed login attempts. Retry-After header is returned. }
...
  /system/lab-validation:
    get:
      summary: Get physical lab validation report.
      description: Admin role required. Summarizes PXE/DHCP/TFTP runtime evidence, BMC endpoint status, SSH access status, safe readiness checks, recent validation runs, and recently recorded physical validation evidence for physical lab acceptance.
      responses:
        '200':
          description: Lab validation report.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LabValidationReport' }
        '403': { description: Admin role required. }
  /system/lab-validation/runs/{id}:
    get:
      summary: Get a persisted physical lab validation run.
      description: Admin role required. Returns one persisted validation run batch and its per-check results for audit and replay review.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: Lab validation run detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LabValidationRunDetail' }
        '400': { description: Invalid run id. }
        '403': { description: Admin role required. }
        '404': { description: Lab validation run not found. }
  /system/lab-validation/runs/{id}/evidence-bundle:
    get:
      summary: Export a physical lab validation evidence bundle.
      description: Admin role required. Returns one validation run, result rows, environment snapshot, readiness checks, related PXE BootEvents, BMC endpoints, SSH accesses, terminal transcripts, script executions, log events, recent physical evidence, a machine-readable operator checklist, and configuration/runtime issues. The bundle is read-only audit material and does not create physical evidence by itself.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: Lab validation evidence bundle.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LabValidationEvidenceBundle' }
        '400': { description: Invalid run id. }
        '403': { description: Admin role required. }
        '404': { description: Lab validation run not found. }
...
  /system/lab-validation/run:
    post:
      summary: Run safe physical lab connectivity checks.
      description: >-
        Admin role and X-Confirm-Action system.lab-validation.run are required.
        Runs read-only PXE HTTP/DHCP/TFTP probes, BMC connectivity checks for
        configured BMC endpoints, BMC tooling checks, and SSH lightweight checks
        for configured endpoints up to the requested limit. Synthetic DHCP probes
        do not create BootEvent evidence, and the HTTP iPXE probe sends
        X-Lab-Validation-Probe so it is recorded as api_event instead of physical
        http_ipxe evidence; pass pxe_macs to verify physical PXE client boot
        events. Strict PXE validation accepts fresh http_ipxe or pxe_dhcp events
        from the last 7 days, and when server_ids are also supplied each PXE
        BootEvent must either carry one of those server_ids or match a requested
        server primary_mac. The target matrix prefers the latest physical PXE
        event over later api_event troubleshooting records. Set strict=true for a
        physical acceptance gate that requires pxe_macs and server_ids for
        targets. BMC is optional per target: a non-simulated BMC adapter is
        required only when requested targets have BMC endpoints; targets without
        BMC endpoints can complete with PXE and SSH proof. Production SSH
        validation still requires SSH_HOST_KEY_POLICY=known_hosts with a readable
        SSH_KNOWN_HOSTS_PATH. It does not perform power operations or enable
        DHCP/TFTP listeners.
      parameters:
        - name: X-Confirm-Action
          in: header
          required: true
          schema: { type: string, enum: [system.lab-validation.run] }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LabValidationRunRequest' }
      responses:
        '200':
          description: Updated lab validation report.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LabValidationReport' }
        '428': { description: Confirmation header is missing. }
...
  /users/{id}:
    patch:
      summary: Update user name or role.
      description: Admin role required. Role must be admin, operator, or viewer. Demoting the last admin user is rejected.
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        '200': { description: User updated. }
        '400': { description: Invalid role. }
        '409': { description: Cannot demote the last admin user. }
...
  /servers/{id}/bom:
    get:
      summary: Get server BOM JSON.
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        '200': { description: BOM payload. }
...